feat(dashboard): nest settings under team
