+++ apps/dashboard/app/routes/teams/$orgId/account/use-settings-data.ts
@@ -4,7 +4,7 @@ import {SettingsLoaderData} from './_loader.server';
 
 export function useSettingsData() {
   const {data: currentUser} = useTypedRouteLoaderData<SettingsLoaderData>(
-    'routes/settings/_route'
+    'routes/teams/$orgId/account/_route'
   )!;
 
   return {currentUser};
