+++ apps/dashboard/app/routes/teams/index/_route.tsx
@@ -1,11 +1,6 @@
 import type {MetaFunction} from '@remix-run/node';
-import {useTypedLoaderData} from 'remix-typedjson';
 
-import {BaseLayout} from '@/components/base-layout';
 import {ErrorPage} from '@/components/error-page';
-
-import {IndexLoaderData} from './_loader.server';
-import {AllTeams} from './index-page';
 
 export {loader} from './_loader.server';
 
@@ -18,18 +13,4 @@ export const meta: MetaFunction = () => {
 
 export function ErrorBoundary() {
   return <ErrorPage />;
-}
-
-export default function IndexPage() {
-  const response = useTypedLoaderData<IndexLoaderData>();
-
-  const {
-    data: {currentUser},
-  } = response;
-
-  return (
-    <BaseLayout title="All Teams" currentUser={currentUser}>
-      <AllTeams currentUser={currentUser} />
-    </BaseLayout>
-  );
 }
